extract load_balancer.go to parts
