+++ pkg/cloud-sdk/blb/backendserver.go
@@ -160,7 +160,7 @@ func (args *UpdateBackendServersArgs) validate() error {
 	return nil
 }
 
-// UpdateBackendServers update  BackendServers
+// UpdateBackendServers update BackendServers
 func (c *Client) UpdateBackendServers(args *UpdateBackendServersArgs) error {
 	err := args.validate()
 	if err != nil {
